Implmented endpoints for Show, Screen, Theatre

+++ src/main/java/com/bookit/backend/model/Show.java
@@ -46,7 +46,7 @@ public class Show {
 
     @NotNull
     @Positive
-    private BigDecimal basePrice;
+    private Double basePrice;
 
 
     @Enumerated(EnumType.STRING)
